# Enhance: Parallelize NOVA layer processing