- Remove blocking Kafka producer.flush() from individual layer publishes
- Replace with non-blocking producer.poll(0) for callbacks
- Move producer flushes to end of parallel processing
- Remove brain wave analogies and simplify timing targets:
  - Reactive: 50ms
  - Responsive: 200ms
  - Reflective: 400ms

Total processing time reduced from ~2s to ~400ms by eliminating
sequential processing bottlenecks. Each layer now processes
independently in parallel, with total time determined by the
slowest layer rather than sum of all layers.

## Summary by Sourcery

Optimize Kafka message publishing by removing blocking operations and moving flushes to the end of processing. Simplify timing targets for processing layers, significantly reducing total processing time by enabling parallel processing.

Enhancements:
- Optimize Kafka message publishing by replacing blocking producer.flush() with non-blocking producer.poll(0) for callbacks, and moving producer flushes to the end of parallel processing.
- Simplify and update timing targets for different processing layers, reducing total processing time from ~2s to ~400ms by eliminating sequential processing bottlenecks.